# apps/backend/src/assets/openapi.yaml
openapi: 3.1.0
info:
  title: Jobly AI API
  version: 0.1.0
  description: |
    OpenAPI contract for Jobly AI — smart job matching & career assistant.
    Conventions:
    - All responses follow the envelope with fields: success, code, message, data?, error?
    - Use bearer JWT auth unless noted; service-to-service calls can use x-api-key.
servers:
  - url: http://localhost:3000/api
    description: Local
  - url: https://api.jobly.ai/v1
    description: Production
  - url: https://staging-api.jobly.ai/v1
    description: Staging
security:
  - bearerAuth: []
...
paths:
...
  /candidates/me:
    get:
      tags: [Candidates]
      summary: Get current candidate profile
      responses:
        '200':
          description: Candidate profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    patch:
      tags: [Candidates]
      summary: Update candidate profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CandidateProfileUpdateRequest'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateProfileResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /candidates/me/resume:
    post:
      tags: [Candidates]
      summary: Upload resume for analysis
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Resume analyzed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CVAnalysisResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /jobs:
    get:
      tags: [Jobs]
      summary: List jobs with filters
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/SortParam'
        - name: q
          in: query
          schema:
            type: string
          description: Free text search
        - name: location
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [full-time, part-time, contract, internship]
        - name: remote
          in: query
          schema:
            type: boolean
        - name: skills
          in: query
          schema:
            type: array
            items:
              type: string
          style: form
          explode: false
      responses:
        '200':
          description: Paginated jobs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobListResponse'
    post:
      tags: [Jobs]
      summary: Create a job (employer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobCreateRequest'
      responses:
        '201':
          description: Job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /jobs/{id}:
    get:
      tags: [Jobs]
      summary: Get job by id
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Job detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Jobs]
      summary: Update job by id
      parameters:
        - $ref: '#/components/parameters/JobId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobUpdateRequest'
      responses:
        '200':
          description: Job updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Jobs]
      summary: Delete job by id
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '204':
          description: Job deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: x-api-key
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    SortParam:
      name: sort
      in: query
      schema:
        type: string
      description: Comma separated field:direction pairs
    JobId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    BaseResponse:
      type: object
      required: [success, code, message]
      properties:
        success:
          type: boolean
        code:
          type: string
          description: Stable application code (e.g., OK, VALIDATION_ERROR)
        message:
          type: string
        traceId:
          type: string
          description: Correlates errors to logs
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          required: [success, code, message]
          properties:
            success:
              type: boolean
              example: false
            error:
              type: object
              properties:
                details:
                  description: Arbitrary validation or domain error details
                traceId:
                  type: string
    PaginatedMeta:
      type: object
      required: [page, pageSize, totalItems, totalPages]
      properties:
        page:
          type: integer
        pageSize:
          type: integer
        totalItems:
          type: integer
        totalPages:
          type: integer
...
    CandidateProfile:
      type: object
      required: [id, userId]
      properties:
        id:
          type: string
        userId:
          type: string
        headline:
          type: string
        summary:
          type: string
        skills:
          type: array
          items:
            type: string
        experience:
          type: array
          items:
            $ref: '#/components/schemas/WorkExperience'
        education:
          type: array
          items:
            $ref: '#/components/schemas/Education'
        links:
          type: array
          items:
            type: string
            format: uri
        resumeUrl:
          type: string
          format: uri
        visibility:
          type: string
          enum: [public, private, connections]
    WorkExperience:
      type: object
      properties:
        title:
          type: string
        company:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
          nullable: true
        description:
          type: string
    Education:
      type: object
      properties:
        school:
          type: string
        degree:
          type: string
        field:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
          nullable: true
...
    Job:
      type: object
      required: [id, employerId, title, description, type, status, createdAt]
      properties:
        id:
          type: string
        employerId:
          type: string
        title:
          type: string
        description:
          type: string
        requirements:
          type: array
          items:
            type: string
        skills:
          type: array
          items:
            type: string
        salaryRange:
          $ref: '#/components/schemas/SalaryRange'
        location:
          type: string
        type:
          type: string
          enum: [full-time, part-time, contract, internship]
        remote:
          type: boolean
        status:
          type: string
          enum: [draft, published, closed]
        createdAt:
          type: string
          format: date-time
    SalaryRange:
      type: object
      properties:
        min:
          type: number
        max:
          type: number
        currency:
          type: string
          example: USD
        period:
          type: string
          enum: [year, month, hour]
...
    CVAnalysis:
      type: object
      properties:
        id:
          type: string
        candidateId:
          type: string
        uploadId:
          type: string
        score:
          type: number
        strengths:
          type: array
          items:
            type: string
        weaknesses:
          type: array
          items:
            type: string
        suggestions:
          type: array
          items:
            type: string
        missingSkills:
          type: array
          items:
            type: string
        readability:
          type: number
          description: Readability score 0-1
        sections:
          type: array
          items:
            $ref: '#/components/schemas/CVSection'
    CVSection:
      type: object
      properties:
        name:
          type: string
        score:
          type: number
        feedback:
          type: string
...
    CandidateProfileUpdateRequest:
      type: object
      properties:
        headline:
          type: string
        summary:
          type: string
        skills:
          type: array
          items:
            type: string
        experience:
          type: array
          items:
            $ref: '#/components/schemas/WorkExperience'
        education:
          type: array
          items:
            $ref: '#/components/schemas/Education'
        links:
          type: array
          items:
            type: string
            format: uri
        visibility:
          type: string
          enum: [public, private, connections]
    CandidateProfileResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            success:
              type: boolean
              example: true
            code:
              type: string
              example: OK
            message:
              type: string
              example: Profile retrieved
            data:
              $ref: '#/components/schemas/CandidateProfile'
...
    JobCreateRequest:
      type: object
      required: [title, description, type]
      properties:
        title:
          type: string
        description:
          type: string
        requirements:
          type: array
          items:
            type: string
        skills:
          type: array
          items:
            type: string
        salaryRange:
          $ref: '#/components/schemas/SalaryRange'
        location:
          type: string
        type:
          type: string
          enum: [full-time, part-time, contract, internship]
        remote:
          type: boolean
    JobUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/JobCreateRequest'
    JobResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            success:
              type: boolean
              example: true
            code:
              type: string
              example: OK
            message:
              type: string
              example: Job retrieved
            data:
              $ref: '#/components/schemas/Job'
    JobListResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            success:
              type: boolean
              example: true
            code:
              type: string
              example: OK
            message:
              type: string
              example: Jobs retrieved
            data:
              type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/Job'
                meta:
                  $ref: '#/components/schemas/PaginatedMeta'
...
    CVAnalysisResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            success:
              type: boolean
              example: true
            code:
              type: string
              example: OK
            message:
              type: string
              example: CV analyzed
            data:
              $ref: '#/components/schemas/CVAnalysis'
...